Moved CodeScannerFragment to fragments package
