+++ collect_app/src/main/java/org/odk/collect/android/preferences/qr/QRCodeTabsAdapter.java
@@ -5,6 +5,8 @@ import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentActivity;
 import androidx.viewpager2.adapter.FragmentStateAdapter;
 
+import org.odk.collect.android.fragments.CodeScannerFragment;
+
 public class QRCodeTabsAdapter extends FragmentStateAdapter {
     public QRCodeTabsAdapter(FragmentActivity fa) {
         super(fa);
