+++ collect_app/src/main/res/layout/fragment_scan.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="org.odk.collect.android.preferences.qr.CodeScannerFragment">
+    tools:context="org.odk.collect.android.fragments.CodeScannerFragment">
 
 
     <com.journeyapps.barcodescanner.DecoratedBarcodeView
